+++ examples/server.js
@@ -1,7 +1,7 @@
 /*
  * Node.js module radiohead-serial
  *
- * Copyright (c) 2017 Peter Müller <peter@crycode.de> (https://crycode.de/)
+ * Copyright (c) 2017-2019 Peter Müller <peter@crycode.de> (https://crycode.de/)
  *
  * Node.js module for communication between some RadioHead nodes and Node.js using
  * the RH_Serial driver and the RHReliableDatagram manager of the RadioHead library.
@@ -11,7 +11,7 @@
  * Copyright (c) 2014 Mike McCauley
  *
  * Port from native C/C++ code to TypeScript
- * Copyright (c) 2017 Peter Müller <peter@crycode.de> (https://crycode.de/)
+ * Copyright (c) 2017-2019 Peter Müller <peter@crycode.de> (https://crycode.de/)
  *
  * Example for the communiation between two nodes.
  * This server example is listening for messages from a client and responding to it.
@@ -22,8 +22,13 @@
 var RadioHeadSerial = require('../').RadioHeadSerial;
 
 // Create an instance of the RadioHeadSerial class
-var rhs = new RadioHeadSerial('/dev/ttyUSB0', 9600, 0x01);
-//var rhs = new RadioHeadSerial('COM1', 9600, 0x01);
+var rhs = new RadioHeadSerial({
+  port: '/dev/ttyUSB0', // COM1 on Windows
+  baud: 9600,
+  address: 0x01,
+  // reliable: true,
+  // autoInit: true
+});
 
 // Listen to the 'data' event for received messages
 rhs.on('data', function(message){